Playing five rounds.
Added win, lose and tie variable.
Changed the console.log under the else, else if statement to return to tie in with the new variables.
Removed the function game, move playerSelection and computerSelection from within a function.
The game is playing five rounds now however the score isn't changing at the minute.